update
tách các api và hàm ra riêng
