+++ controllers/thongke/thong-ke-controller.js
@@ -282,87 +282,13 @@ const thongKeDoanhThuTheoThangTrongNam = async (req, res, next) => {
         };
     }
 };
+// Thông kê món
 
-
-// api 
-const thongKeDoanhThuTheoNgayApi = async (req, res, next) => {
-    try {
-        const result = await thongKeDoanhThuTheoNgay(req, res, next);
-        res.status(200).json(result);
-        // Send the result directly without using JSON.stringify
-
-    } catch (error) {
-        // Check if headers have already been sent
-        if (res.headersSent) {
-            console.error(" Tiêu đề đã được gửi đi rồi. Không thể gửi phản hồi lỗi.");
-        } else {
-            res.status(500).json({ msg: 'Đã xảy ra lỗi khi kích hoạt hóa đơn', error: error.message });
-        }
-    }
-}
-const thongKeDoanhThuTheo10NgayApi = async (req, res, next) => {
-    try {
-        const result = await thongKeDoanhThuTheo10Ngay(req, res, next);
-        res.status(200).json(result);  // Send the result directly without using JSON.stringify
-    } catch (error) {
-        // Check if headers have already been sent
-        if (res.headersSent) {
-            console.error(" Tiêu đề đã được gửi đi rồi. Không thể gửi phản hồi lỗi.");
-        } else {
-            res.status(500).json({ msg: 'Đã xảy ra lỗi khi kích hoạt hóa đơn', error: error.message });
-        }
-    }
-}
-const thongKeDoanhThuTheo30NgayApi = async (req, res, next) => {
-    try {
-        const result = await thongKeDoanhThuTheo30Ngay(req, res, next);
-        res.status(200).json(result);  // Send the result directly without using JSON.stringify
-    } catch (error) {
-        // Check if headers have already been sent
-        if (res.headersSent) {
-            console.error(" Tiêu đề đã được gửi đi rồi. Không thể gửi phản hồi lỗi.");
-        } else {
-            res.status(500).json({ msg: 'Đã xảy ra lỗi khi kích hoạt hóa đơn', error: error.message });
-        }
-    }
-}
-
-const thongKeDoanhThuTheoNamApi = async (req, res, next) => {
-    try {
-        const result = await thongKeDoanhThuTheoNam(req, res, next);
-        res.status(200).json(result);  // Send the result directly without using JSON.stringify
-    } catch (error) {
-        // Check if headers have already been sent
-        if (res.headersSent) {
-            console.error(" Tiêu đề đã được gửi đi rồi. Không thể gửi phản hồi lỗi.");
-        } else {
-            res.status(500).json({ msg: 'Đã xảy ra lỗi khi kích hoạt hóa đơn', error: error.message });
-        }
-    }
-}
-
-const thongKeDoanhThuTheoThangTrongNamApi = async (req, res, next) => {
-    try {
-        const monthlyRevenue = await thongKeDoanhThuTheoThangTrongNam(req, res, next);
-        res.status(200).json(monthlyRevenue);  // Send the result directly without using JSON.stringify
-    } catch (error) {
-        // Check if headers have already been sent
-        if (res.headersSent) {
-            console.error(" Tiêu đề đã được gửi đi rồi. Không thể gửi phản hồi lỗi.");
-        } else {
-            res.status(500).json({ msg: 'Đã xảy ra lỗi khi kích hoạt hóa đơn', error: error.message });
-        }
-    }
-}
 module.exports = {
     thongKeDoanhThuTheoNgay,
     thongKeDoanhThuTheo10Ngay,
     thongKeDoanhThuTheo30Ngay,
     thongKeDoanhThuTheoNam,
-    thongKeDoanhThuTheoNgayApi,
-    thongKeDoanhThuTheo10NgayApi,
-    thongKeDoanhThuTheo30NgayApi,
-    thongKeDoanhThuTheoNamApi,
-    thongKeDoanhThuTheoThangTrongNam,
-    thongKeDoanhThuTheoThangTrongNamApi
+    thongKeDoanhThuTheoThangTrongNam
+
 }
